Refactor RegisteredApplication management.
Create service for managing RegisteredApplication and created 2 endpoints:
 - one for managing applications within project
 - another one for managing single application
